修复改变弹幕大小和速度弹幕显示不全的问题

+++ JHDanmakuRender/JHDanmakuClock.h
@@ -16,7 +16,7 @@
 @interface JHDanmakuClock : NSObject
 @property (weak, nonatomic) id<JHDanmakuClockDelegate> delegate;
 @property (assign, nonatomic) NSTimeInterval offsetTime;
-@property (assign, nonatomic) float speed;
+@property (assign, nonatomic) CGFloat speed;
 - (void)setCurrentTime:(NSTimeInterval)currentTime;
 - (void)start;
 - (void)stop;
